+++ module/helpers/constants.mjs
@@ -153,13 +153,14 @@ export const DEFAULT_CARAVAN_ICON = 'icons/svg/target.svg';
 
 export const DEFAULT_BURDEN_ICON = 'icons/svg/stoned.svg';
 export const DEFAULT_GEAR_ICON = 'icons/svg/item-bag.svg';
-export const DEFAULT_TRAIT_ICON = 'icons/svg/aura.svg';
+export const DEFAULT_TRAIT_ICON = 'icons/svg/walk.svg';
 export const DEFAULT_ARMOR_ICON = 'icons/svg/shield.svg';
 export const DEFAULT_POWER_ICON = 'icons/svg/fire.svg';
 export const DEFAULT_POWER_ALBUM_ICON = 'icons/svg/book.svg';
-export const DEFAULT_WARD_ICON = 'icons/svg/frozen.svg';
+export const DEFAULT_WARD_ICON = 'icons/svg/eye.svg';
 export const DEFAULT_WEAPON_ICON = 'icons/svg/sword.svg';
 export const DEFAULT_CASH_ICON = 'icons/commodities/currency/coins-stitched-pouch-brown.webp';
+export const DEFAULT_SKILL_ICON = 'icons/svg/jump.svg';
 
 export const GEAR_ICONS = [
   DEFAULT_GEAR_ICON,